Refatora testes do FiltroCard
